fix USES_GPIO and merge it into USES_PIN
It doesn't make sense to claim the gpio without claiming the pin or vice
versa, so merge these macros.

Also implement a gross workaround for the problem of gpio node names
getting mangled because they match the names of pin macros.  The nice
solution would be to enhance dtc to support numerical constants that can
be used in property values, instead of using cpp macros which are
expanded in every context.